+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2] - 2022-06-12 - @damien.clement
+### Changed
+- Move entry point to own header file
+- Implement all core function as static for multiple include
+
 ## [0.1.1] - 2022-06-07 - @damien.clement
 ### Changed
 - Refactor names for parameters from `t_*` `p_*`